Changed files, updated README
Added the script we're using to download and install Twingate on a PC that doesn't have .NET 6 Desktop Runtime installed. Changed the name of the Desktop Runtime file to an arbitrary name so when the files are updated later on, the script doesn't have to rely on the name of the file being an old version... If that makes sense.